Item 3/requirements.txt
pandas == 2.1.4
python-dotenv == 1.0.0
openai == 1.3.7